Version 0.92
Well, just polishing things up.  I am now very confident that it will work.  I ran the command line version on /usr/include and got a syntax highlighting file of 7000 lines.  I included that file in the tar.  I also added some things to aid in expanding this to work with other languages.  For now though it is for C++ only.