Version 1.1
Changed a few things and fixed some bugs.  Mainly, I made it so you do not have to move any files to make the plugin work, if you don't want to.  It is still *highly* recommend that you do move the files as suggested in the help file, but you now have the option to not move anything.  Updated the help files, and fixed some typos and other minor things.